+++ Bitfinex.Net/Objects/BitfinexAccountInfo.cs
@@ -6,9 +6,9 @@ namespace Bitfinex.Net.Objects
     public class BitfinexAccountInfo
     {
         [JsonProperty("maker_fees")]
-        public double MakerFees { get; set; }
+        public decimal MakerFees { get; set; }
         [JsonProperty("taker_fees")]
-        public double TakerFees { get; set; }
+        public decimal TakerFees { get; set; }
         [JsonProperty("fees")]
         public List<BitfinexFee> Fees { get; set; }
     }
@@ -18,8 +18,8 @@ namespace Bitfinex.Net.Objects
         [JsonProperty("pairs")]
         public string Pairs { get; set; }
         [JsonProperty("maker_fees")]
-        public double MakerFees { get; set; }
+        public decimal MakerFees { get; set; }
         [JsonProperty("taker_fees")]
-        public double TakerFees { get; set; }
+        public decimal TakerFees { get; set; }
     }
 }
