+++ Bitfinex.Net/Objects/BitfinexOrderBookEntry.cs
@@ -7,10 +7,10 @@ namespace Bitfinex.Net.Objects
     public class BitfinexOrderBookEntry
     {
         [BitfinexProperty(0)]
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         [BitfinexProperty(1)]
         public int Count { get; set; }
         [BitfinexProperty(2)]
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
     }
 }
